+++ src/test/integration/lint_test.ts
@@ -16,6 +16,7 @@ import {assert} from 'chai';
 import * as path from 'path';
 import {runCommand} from './run-command';
 import {invertPromise} from '../util';
+import {child_process} from 'mz';
 
 const fixturePath =
     path.join(__dirname, '../../../src/test/integration/fixtures/');
@@ -183,6 +184,77 @@ Fixed 4 warnings.
 </dom-module>
 `);
   });
+
+  suite('--watch', function() {
+
+    this.timeout(12 * 1000);
+
+    const delimiter =
+        `\nLint pass complete, waiting for filesystem changes.\n\n`;
+    test('re-reports lint results when the filesystem changes', async () => {
+      const fixtureDir = path.join(fixturePath, 'lint-simple');
+      const cwd = getTempCopy(fixtureDir);
+      const forkedProcess =
+          child_process.fork(binPath, ['lint', '--watch'], {cwd, silent: true});
+      const reader = new StreamReader(forkedProcess.stdout);
+      assert.deepEqual(await reader.readUntil(delimiter), '');
+      fs.writeFileSync(
+          path.join(cwd, 'my-elem.html'),
+          '<style>\nfoo {@apply(--bar)}\n</style>');
+      assert.deepEqual(await reader.readUntil(delimiter), `
+
+foo {@apply(--bar)}
+           ~~~~~~~
+
+my-elem.html(1,11) error [at-apply-with-parens] - @apply with parentheses is deprecated. Prefer: @apply --foo;
+
+
+Found 1 error. 1 can be automatically fixed with --fix.
+`);
+      // Wait for a moment
+      await sleep(300);
+      // Fix the error
+      fs.writeFileSync(
+          path.join(cwd, 'my-elem.html'),
+          '<style>\nfoo {@apply --bar}\n</style>');
+      // Expect empty output again.
+      assert.deepEqual(await reader.readUntil(delimiter), ``);
+      // Expect no other output.
+      await sleep(200);
+      assert.deepEqual(await reader.readRestOfBuffer(), '');
+      forkedProcess.kill();
+    });
+
+    const testName =
+        'with --fix, reports and fixes when the filesystem changes';
+    test(testName, async () => {
+      const fixtureDir = path.join(fixturePath, 'lint-simple');
+      const cwd = getTempCopy(fixtureDir);
+      const forkedProcess = child_process.fork(
+          binPath, ['lint', '--watch', '--fix'], {cwd, silent: true});
+      const reader = new StreamReader(forkedProcess.stdout);
+      // The first pass yields no warnings:
+      assert.deepEqual(
+          await reader.readUntil(delimiter), 'No fixes to apply.\n');
+      // Add an error
+      fs.writeFileSync(
+          path.join(cwd, 'my-elem.html'),
+          '<style>\nfoo {@apply(--bar)}\n</style>');
+      // Expect warning output.
+      assert.deepEqual(
+          await reader.readUntil(delimiter), `  Made 2 changes to my-elem.html
+
+Fixed 1 warning.
+`);
+      // The automated fix triggers the linter to run again.
+      assert.deepEqual(
+          await reader.readUntil(delimiter), 'No fixes to apply.\n');
+      // Expect no other output.
+      await sleep(200);
+      assert.deepEqual(await reader.readRestOfBuffer(), '');
+      forkedProcess.kill();
+    });
+  });
 });
 
 function getTempCopy(fromDir: string) {
@@ -204,3 +276,42 @@ function copyDir(fromDir: string, toDir: string) {
     }
   }
 };
+
+/** Simple class for reading up to a delimitor in an unending stream. */
+class StreamReader {
+  private buffer = '';
+  private wakeup: () => void = () => undefined;
+  constructor(readStream: NodeJS.ReadableStream) {
+    readStream.setEncoding('utf8');
+    readStream.on('data', (chunk: string) => {
+      this.buffer += chunk;
+      if (this.wakeup) {
+        this.wakeup();
+      }
+    });
+  }
+
+  async readUntil(text: string) {
+    while (true) {
+      const index = this.buffer.indexOf(text);
+      if (index >= 0) {
+        const result = this.buffer.slice(0, index);
+        this.buffer = this.buffer.slice(index + text.length + 1);
+        return result;
+      }
+      await new Promise<void>((resolve) => {
+        this.wakeup = resolve;
+      });
+    }
+  }
+
+  async readRestOfBuffer() {
+    const result = this.buffer;
+    this.buffer = '';
+    return result;
+  }
+}
+
+async function sleep(millis: number) {
+  return new Promise((resolve) => setTimeout(resolve, millis));
+}