+++ CHANGELOG.md
@@ -5,6 +5,8 @@
   - Also supports warnings which can be addressed with less-safe changes via
     an interactive prompt and the `--edits` flag. See `polymer lint --help` for
     more info.
+- Added `--watch` (shorthand `-w`) option to `polymer lint`. When passed, we will watch the filesystem for changes and rerun the linter immediately afterwards.
+  - Also works with `--fix` to automatically fix and report warnings as you work!
 - `build` Added a CLI argument for setting the `basePath` option: `--base-path`.
 <!-- Add new, unreleased items here. -->